Fixed a bug related to DST
There was an off-by-one-hour timezones that have DST, when calculating an ISO8601 date outside the timezone that was in-effect on January 1.

E.g. on Jan 1 in NSW, Australia, the timezone is +11 (due to DST), however on April 30 it is +10. This meant that, as the timezone wasn't being re-computed after the day/month/year was parsed, we could have UTC times incorrectly parsing out to 11am local time when DST was NOT in effect (should be 10am in that instance).